pyproject.toml
[project]
name = "bbai"
version = "0.1.0"
description = "Bug Bounty AI Agent - Production-grade automation framework with visible AI reasoning"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [
    {name = "BBAI Team"}
]
keywords = ["bug-bounty", "security", "ai", "automation", "pentesting"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "Topic :: Security",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.14",
]

dependencies = [
    # Async & Core
    "anyio[trio]>=4.4.0",
    "asyncio-mqtt>=0.16.0",
    "aiofiles>=24.1.0",
    "aioshutil>=1.5",
    
    # CLI & Interface
    "typer>=0.12.3",
    "rich>=13.7.1",
    "prompt-toolkit>=3.0.0",
    "questionary>=2.0.1",
    
    # AI & LangChain
    "langgraph>=0.2.14",
    "langchain-core>=0.3.0",
    "langgraph-checkpoint-sqlite>=1.0.0",
    
    # Data & Validation
    "pydantic>=2.8.2",
    "pydantic-settings>=2.3.4",
    "orjson>=3.10.6",
    "validators>=0.28.0",
    "publicsuffixlist>=0.10.0",
    "tldextract>=5.1.2",
    
    # Storage
    "aiosqlite>=0.20.0",
    "sqlalchemy[asyncio]>=2.0.31",
    
    # Docker & System
    "docker>=7.1.0",
    
    # Security & Rate Limiting
    "aiolimiter>=1.1.0",
    "cryptography>=42.0.0",
    
    # Utilities
    "pyyaml>=6.0.1",
    "pytz>=2024.1",
    "tenacity>=8.5.0",
    "structlog>=24.4.0",
    "python-dotenv>=1.0.1",
    "httpx[http2]>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.2",
    "pytest-asyncio>=0.23.8",
    "pytest-cov>=5.0.0",
    "pytest-xdist>=3.6.1",
    "respx>=0.21.0",
    "black>=24.8.0",
    "ruff>=0.6.0",
    "mypy>=1.11.1",
    "pre-commit>=3.8.0",
    "bandit>=1.7.9",
]

[project.scripts]
bbai = "bbai.cli.main:app"

[project.urls]
Homepage = "https://github.com/bbai/bbai"
Repository = "https://github.com/bbai/bbai"
Documentation = "https://bbai.readthedocs.io"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
resolution = "highest"
prerelease = "allow"

[tool.hatch.build.targets.wheel]
packages = ["src/bbai"]

[tool.black]
line-length = 100
target-version = ['py314']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py314"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "ASYNC", # flake8-async
    "C90", # mccabe complexity
]
ignore = [
    "E501",  # Line too long (handled by black)
]

[tool.ruff.pydocstyle]
convention = "google"

[tool.ruff.mccabe]
max-complexity = 10

[tool.mypy]
python_version = "3.14"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true
show_column_numbers = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers --tb=short"
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["src/bbai"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
show_missing = true
skip_covered = false
